requirements.txt
opencv-python==4.8.1.78
matplotlib==3.7.2
numpy==1.24.3
Pillow==10.0.0
